refactor(ws): rename infra.rs to messages.rs
